Add CenteredStepLayout component
