+++ packages/onboarding/src/step-wireframes/centered-layout/CenteredStepLayout.stories.tsx
@@ -1,0 +1,102 @@
+import SelectCardCheckbox from '../../select-card-checkbox';
+import { CenteredStepLayout } from './CenteredStepLayout';
+import type { Meta } from '@storybook/react';
+
+import './style.stories.scss';
+
+const meta: Meta< typeof CenteredStepLayout > = {
+	title: 'Onboarding/StepWireframes/CenteredStepLayout',
+	component: CenteredStepLayout,
+};
+
+export default meta;
+
+const goals = [
+	{
+		key: 'write',
+		title: 'Publish a blog',
+	},
+	{
+		key: 'engagement',
+		title: 'Build and engage an audience',
+	},
+	{
+		key: 'collect-donations',
+		title: 'Collect donations',
+	},
+	{
+		key: 'portfolio',
+		title: 'Showcase work/portfolio',
+	},
+	{
+		key: 'build-nonprofit',
+		title: 'Build a site for a school or nonprofit',
+	},
+	{
+		key: 'newsletter',
+		title: 'Create a newsletter',
+	},
+	{
+		key: 'sell-digital',
+		title: 'Sell services or digital goods',
+	},
+	{
+		key: 'sell-physical',
+		title: 'Sell physical goods',
+	},
+	{
+		key: 'promote',
+		title: 'Promote my business',
+	},
+	{
+		key: 'courses',
+		title: 'Create a course',
+	},
+	{
+		key: 'contact-form',
+		title: 'Create a contact form',
+	},
+	{
+		key: 'videos',
+		title: 'Upload videos',
+	},
+	{
+		key: 'paid-subscribers',
+		title: 'Offer paid content to members',
+	},
+	{
+		key: 'announce-events',
+		title: 'Announce events',
+	},
+];
+
+export const Goals = () => {
+	return (
+		<CenteredStepLayout
+			heading={ {
+				text: 'What would you like to create?',
+				subText: 'Pick one or more goals to get started.',
+			} }
+			backButton={ () => {} }
+			skipButton={ () => {} }
+			nextButton={ {
+				label: 'Continue',
+				onClick: () => {},
+			} }
+			render={ ( { nextButton } ) => {
+				return (
+					<>
+						<div className="goals__list">
+							{ goals.map( ( goal ) => (
+								<SelectCardCheckbox key={ goal.key } checked={ false } onChange={ () => {} }>
+									<span>{ goal.title }</span>
+								</SelectCardCheckbox>
+							) ) }
+						</div>
+						{ nextButton && <div style={ { marginTop: '3rem' } }>{ nextButton }</div> }
+					</>
+				);
+			} }
+		/>
+	);
+};
