+++ packages/onboarding/src/step-wireframes/centered-layout/style.scss
@@ -1,0 +1,4 @@
+.centered-step-layout {
+	display: grid;
+	justify-items: center;
+}
