+++ packages/onboarding/src/step-wireframes/index.ts
@@ -1,1 +1,2 @@
 export { HorizontalStepLayout } from './horizontal-layout/HorizontalStepLayout';
+export { CenteredStepLayout } from './centered-layout/CenteredStepLayout';
